Support OpenBSD
Recent versions of OpenBSD use LLD, the LLVM linker, by default.  This
linker is compatible with GNU linkers, so this commit changes the
compiler_description entry for LD so that its version regexp matches
LLD's version string.

The other linker configurations are like those used on Linux and
FreeBSD, so I added OpenBSD to the relevant target patterns.

Using this configuration, I've been able to build a wide array of Ada
tools using gprconfig and gprbuild.