Use -mno-outline-atomics on aarch64-vx7r2cert
__aarch64_swp1_acq is provided by libgcc, but we want to avoid that
dependency since we don't want users to link with libgcc when using
restricted runtimes (this would introduce issues with certification).

Note that the previous change was introducing a regression in
kb-validation because of two non-contiguous nodes "compilers" in the
same configuration.  this old change has been reverted; the problem is
fixed in this new change by changing the order of nodes to make them
contiguous.

TN: UA18-045
Change-Id: Ia85396bdf49b8730d4a05c8ba4da191c9156c8a7