U928-033 compilers.xml: remove default rts-device-cuda directory
We do not want gprconfig to default to a runtime and instead want to
force the usage of the runtime specified through gprfiles. This fixes an
issue where the runtime would be selected by reading the outmut of
llvm-gnatls, which would point to the native runtime instead of the
device one.

Change-Id: Ibb54e6c7df16c962804e4303cac792c5821ac193