add support for arm-qnx (V303-022)
There is no arm-qnx compiler available so far that we can use
to validate this change, but they just reproduce for arm-qnx
the configuration that we have for aarch64-qnx. Just out of
caution, I verified that I was able to rebuild gprbuild, and then
re-run the gprbuild testing on aarch64-qnx without problem.

Change-Id: Ic09459a9936345f85eb77ea1b3088db2f22684e4