Many Linux distros do not give the full version in the gcc directory structure, ie /usr/lib64/gcc/x86_64-pc-linux/7 rather than /usr/lib64/gcc/x86_64-pc-linux/7.5.0, limiting it to only the major verion number rather than the full version number. We therefore should check for both.
On 64bit systems with a 32bit underlay, you cannot have the dual files
under the same directory. It is normal therefore to have 32bit systems
under .../lib and 64bit systems under .../lib64 and others. However,
gprconfig only checks .../lib. We therefore check lib, lib32 & lib64 to
ensure the Ada runtime is picked up. As the regex has a hefty path to
check, the overhead that this produces is minimal.