Prevent -z origin in the Run_Path_Option for Android
The dynamic linker on old versions of Android raises warnings
on the executable flag set by ld -z origin, claiming it does
not know about it.

$ORIGIN gets processed regardless, as the change acknowledging
the flag in more recent versions suggests:

    commit 7aa3abc781e37d3ba5a3642ab6379d83d8aab1a3

    Add DF_1_ORIGIN to SUPPORTED_DT_FLAGS_1.

    This flag means "$ORIGIN processing required", and since we always
    do that, we can claim support for it.

The change proposed here prevents the warning by setting up a separate
RUNPATH processing block for Android, similar to the linux one but for
Run_Path_Option. -z origin is simply omitted to prevent the warning.